# feat(core): Docusaurus Faster - Rspack Persistent Cache