--- a/packages/docusaurus-plugin-content-docs/src/index.ts
+++ b/packages/docusaurus-plugin-content-docs/src/index.ts
@@ -76,7 +76,14 @@ async function createMdxLoaderDependencyFile({
   dataDir: string;
   options: PluginOptions;
   versionsMetadata: VersionMetadata[];
-}) {
+}): Promise<string | undefined> {
+  // TODO this has been temporarily made opt-in until Rspack cache bug is fixed
+  //  See https://github.com/facebook/docusaurus/pull/10931
+  //  See https://github.com/facebook/docusaurus/pull/10934#issuecomment-2672253145
+  if (!process.env.DOCUSAURUS_ENABLE_MDX_DEPENDENCY_FILE) {
+    return undefined;
+  }
+
   const filePath = path.join(dataDir, '__mdx-loader-dependency.json');
   // the cache is invalidated whenever this file content changes
   const fileContent = {
@@ -138,7 +145,7 @@ export default async function pluginContentDocs(
             options,
             versionsMetadata,
           }),
-        ],
+        ].filter((d): d is string => typeof d === 'string'),
 
         useCrossCompilerCache:
           siteConfig.future.experimental_faster.mdxCrossCompilerCache,