--- a/packages/docusaurus/src/webpack/base.ts
+++ b/packages/docusaurus/src/webpack/base.ts
@@ -35,9 +35,19 @@ const LibrariesToTranspileRegex = new RegExp(
   LibrariesToTranspile.map((libName) => `(node_modules/${libName})`).join('|'),
 );
 
+// TODO later: centralize/validate all env variables in a single place
+const Env = {
+  isProd: process.env.NODE_ENV === 'production',
+
+  // Secret flags to disable features that may cause troubles
+  noReactAliases: !!process.env.DOCUSAURUS_NO_REACT_ALIASES,
+  noPersistentCache: !!process.env.DOCUSAURUS_NO_PERSISTENT_CACHE,
+  noRspackIncremental: !!process.env.DOCUSAURUS_NO_RSPACK_INCREMENTAL,
+};
+
 function getReactAliases(siteDir: string): Record<string, string> {
   // Escape hatch
-  if (process.env.DOCUSAURUS_NO_REACT_ALIASES) {
+  if (Env.noReactAliases) {
     return {};
   }
   const resolveSitePkg = (id: string) =>
@@ -87,13 +97,12 @@ export async function createBaseConfig({
     plugins,
   } = props;
   const totalPages = routesPaths.length;
-  const isProd = process.env.NODE_ENV === 'production';
-  const minimizeEnabled = minify && isProd;
+  const minimizeEnabled = minify && Env.isProd;
 
   const fileLoaderUtils = getFileLoaderUtils(isServer);
 
   const name = isServer ? 'server' : 'client';
-  const mode = isProd ? 'production' : 'development';
+  const mode = Env.isProd ? 'production' : 'development';
 
   const themeAliases = await loadThemeAliases({siteDir, plugins});
 
@@ -150,7 +159,7 @@ export async function createBaseConfig({
         // See https://github.com/web-infra-dev/rspress/pull/1631
         // See https://github.com/facebook/docusaurus/issues/10646
         // @ts-expect-error: Rspack-only, not available in Webpack typedefs
-        incremental: !isProd && !process.env.DISABLE_RSPACK_INCREMENTAL,
+        incremental: !isProd && !Env.noRspackIncremental,
       };
     }
     return undefined;
@@ -164,8 +173,10 @@ export async function createBaseConfig({
     output: {
       pathinfo: false,
       path: outDir,
-      filename: isProd ? 'assets/js/[name].[contenthash:8].js' : '[name].js',
-      chunkFilename: isProd
+      filename: Env.isProd
+        ? 'assets/js/[name].[contenthash:8].js'
+        : '[name].js',
+      chunkFilename: Env.isProd
         ? 'assets/js/[name].[contenthash:8].js'
         : '[name].js',
       publicPath:
@@ -176,7 +187,7 @@ export async function createBaseConfig({
     performance: {
       hints: false,
     },
-    devtool: isProd ? undefined : 'eval-cheap-module-source-map',
+    devtool: Env.isProd ? undefined : 'eval-cheap-module-source-map',
     resolve: {
       extensions: ['.wasm', '.mjs', '.js', '.jsx', '.ts', '.tsx', '.json'],
       symlinks: true, // See https://github.com/facebook/docusaurus/issues/3272
@@ -270,7 +281,7 @@ export async function createBaseConfig({
           exclude: CSS_MODULE_REGEX,
           use: configureWebpackUtils.getStyleLoaders(isServer, {
             importLoaders: 1,
-            sourceMap: !isProd,
+            sourceMap: !Env.isProd,
           }),
         },
         // Adds support for CSS Modules (https://github.com/css-modules/css-modules)
@@ -285,17 +296,17 @@ export async function createBaseConfig({
               exportOnlyLocals: isServer,
             },
             importLoaders: 1,
-            sourceMap: !isProd,
+            sourceMap: !Env.isProd,
           }),
         },
       ],
     },
     plugins: [
       new CSSExtractPlugin({
-        filename: isProd
+        filename: Env.isProd
           ? 'assets/css/[name].[contenthash:8].css'
           : '[name].css',
-        chunkFilename: isProd
+        chunkFilename: Env.isProd
           ? 'assets/css/[name].[contenthash:8].css'
           : '[name].css',
         // Remove css order warnings if css imports are not sorted