--- a/packages/docusaurus/src/webpack/base.ts
+++ b/packages/docusaurus/src/webpack/base.ts
@@ -147,6 +147,7 @@ export async function createBaseConfig({
         return disabledPersistentCacheValue;
       }
     }
+
     return {
       type: 'filesystem',
       name: getCacheName(),